Authentication layer WIP

+++ ProgrammingUnit/Furniture.h
@@ -10,7 +10,7 @@ protected:
 	string name;
 	float basePrice;
 public:
-	virtual int calculatePrice(int quantity) = 0;
+	virtual float calculatePrice(int quantity) = 0;
 protected:
 	void setName(string name);
 	void setBasePrice(float pr);
