+++ ProgrammingUnit/Table.cpp
@@ -6,6 +6,6 @@ Table::Table(string name, float price) {
 	setBasePrice(price);
 	setName(name);
 }
-int Table::calculatePrice(int quantity = 1) {
+float Table::calculatePrice(int quantity = 1) {
 	return basePrice * quantity;
 }
